requirement.txt
pytest>=7.0.0
pytest-html>=3.1.0
pytest-xdist>=2.5.0
pytest-timeout>=2.1.0
pytest-rerunfailures>=10.2
paramiko>=2.9.0
scapy>=2.4.5
pyyaml>=6.0
pandas>=1.3.0
jinja2>=3.0.0